refactor(scripts): split session_report.py's flat CLI into per-profile modules
session_report.py was a single 2,269-line file mixing argument parsing with
both profiles' extraction logic, and extract_downstream()/combine() had
internally duplicated rework/gate/utilization/accuracy JSON-shaping blocks.

Split into scripts/lib/session_report_shared.py (building blocks both
profiles need), session_report_maintainer.py, and
session_report_downstream.py; session_report.py is now a 486-line thin CLI
dispatcher that re-exports every name tests and external consumers
(scripts/bash_failure_taxonomy.py, scripts/eval_rawlog.py) previously
imported from it, so the CLI surface and every downstream/maintainer digest
byte-for-byte match the pre-split golden fixtures. The downstream profile's
duplicated shaping is now four shared _shape_* helpers.

Fixes #2098.
